+++ package.json
@@ -11,10 +11,12 @@
     "@types/node": "^16.18.95",
     "@types/react": "^18.2.75",
     "@types/react-dom": "^18.2.24",
+    "axios": "^1.6.8",
     "bootstrap": "^5.3.3",
     "font-awesome": "^4.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-dropzone": "^14.2.3",
     "react-icons": "^5.1.0",
     "react-redux": "^9.1.0",
     "react-router": "^6.22.3",
